+++ src/main/resources/templates/citydata/citydatalist.html
@@ -7,7 +7,11 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>WhatSeoul</title>
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css">
-    <link rel="stylesheet" type="text/css" href="../../static/css/citydata/citydatalist.css" th:href="@{../../static/css/index/index.css}">
+    <!-- 도시데이터 전체 조회(이 페이지)를 위한 css : 내부에 인구데이터 관련 css는 인구데이터 관련이라고 표기해둠    -->
+    <link rel="stylesheet" type="text/css" href="/css/citydata/citydatalist.css" th:href="@{/css/citydata/citydatalist.css}">
+    <!-- 인구데이터 조회 기능을 위한 js -->
+    <script defer src="/js/citydata/population.js"></script>
+
 </head>
 <body>
 <header>
@@ -45,17 +49,38 @@
 <main>
     <section class="main-content-wrapper">
         <div class="main-content">
-            <div class="main-content-title">
-                지금 서울은?
-            </div>
+            <!--${areaName}은 지금...-->
+            <div class="main-content-title"></div>
             <div class="main-content-descrip">
-                서울시내 실시간 <strong>인구</strong>, <strong>날씨</strong>, <strong>문화행사</strong> 정보를 제공합니다.<br>
+                <!--서울시내 실시간 <strong>인구</strong>, <strong>날씨</strong>, <strong>문화행사</strong> 정보를 제공합니다.<br>-->
             </div>
-            <img alt="실시간 지역정보 제공 기능 설명 이미지" class="blank-img" src="../../static/media/blank_img.png"/>
+            <!-- 실시간 날씨 현황 -->
+
+            <!-- 실시간 인구 현황 -->
+            <div class="main-content-population-wrapper">
+                <div class="main-content-population">
+                    <strong style="font-size: 20px">실시간 인구 현황</strong><br>
+                    실시간 인구 혼잡도 : <span id="area-congest-lvl"></span><br>
+                    <span id="area-congest-msg" style="color: gray"></span>
+                </div><br>
+                <div class="main-content-population-forecast">
+                    <!--인구 예측지표 차트가 그려질 부분-->
+                    장소 예측 인구
+                    <canvas id="myChart"></canvas>
+                </div>
+            </div>
+
+            <!-- 실시간 문화행사 정보 -->
+
+
+
         </div>
+
     </section>
+
+    <!-- 계속 떠다니는 버튼들 필요없을시 제거  -->
     <section class="buttons-wrapper" >
-        <button class="location-info-btn" onclick="location.href='../selectarea/selectdistrict.html'">
+        <button class="location-info-btn" onclick="location.href='/district'">
             실시간<br>
             지역정보 확인하기</button>
         <button class="alan-question-btn" href="#">
@@ -65,14 +90,16 @@
                     서울 명소 추천받기
                 </div>
                 <div class="btn-contet-img">
-                    <img alt="앨런 AI 아이콘" class="alan-question-btn-img" src="../../static/media/alan_logo_no_bg.png"/>
+                    <img alt="앨런 AI 아이콘" class="alan-question-btn-img" src="/media/alan_logo_no_bg.png"/>
                 </div>
             </div>
-
         </button>
     </section>
+    <!-- 계속 떠다니는 버튼들 필요없을시 제거  -->
 
 </main>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.4.0/Chart.min.js"></script>
+
 <!-- 부트스트랩 JS (jQuery 포함) -->
 <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js"></script>
