+++ src/main/resources/templates/index/index.html
@@ -83,7 +83,7 @@
         </div>
     </section>
     <section class="buttons-wrapper" >
-        <button class="location-info-btn" onclick="location.href='../selectarea/selectdistrict.html'">
+        <button class="location-info-btn" onclick="location.href='/district'">
             실시간<br>
             지역정보 확인하기</button>
         <button class="alan-question-btn" href="#">
