+++ src/main/resources/templates/selectarea/selectconfirm.html
@@ -50,7 +50,7 @@
             해당 지역의 실시간 지역정보를 조회하겠습니다.
             <div class="button-container-wrapper">
                 <div class="button-container">
-                    <button id="confirmButton" onclick="location.href='#'">네, 좋아요</button>
+                    <button id="confirmButton" onclick="location.href='/citydata'">네, 좋아요</button>
                     <button id="backButton" onclick="location.href='/area'">뒤로가기</button>
                 </div>
             </div>
